requirements.txt
aiogram==2.24
python-dotenv==0.21.1
